LayoutTests/accessibility/label-for-with-aria-labelledby-expected.txt

This tests that the aria-labelledby pointing to a label works when the label has a `for` attribute.

	AXTitle: Hello
	AXDescription: Hello
	AXHelp:
Changing inner HTML of label:
	AXTitle: World
	AXDescription: World
	AXHelp:

PASS successfullyParsed is true

TEST COMPLETE
 World
